+++ packages/tonal/CHANGELOG.md
@@ -1,5 +1,19 @@
 # tonal
 
+## 4.10.0
+
+### Minor Changes
+
+- Scale.degrees new function
+
+  Example: `[1, 2, 3].map(Scale.degrees("C major")) => ["C", "D", "E"]`
+
+### Patch Changes
+
+- Updated dependencies
+  - @tonaljs/scale@4.9.0
+  - @tonaljs/note@4.9.0
+
 ## 4.9.0
 
 ### Minor Changes